% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buat_laporan.R
\name{render_laporan}
\alias{render_laporan}
\title{Render proyek laporan menggunakan Quarto}
\usage{
render_laporan(path = ".")
}
\arguments{
\item{path}{Direktori proyek laporan.}
}
\value{
(Secara \emph{invisible}) \code{path}.
}
\description{
Pembungkus tipis untuk menjalankan \verb{quarto render} pada direktori proyek.
}
